requirements.txt
langchain>=0.0.325
langchain-community>=0.0.13
langchain-openai>=0.0.2
langchain-huggingface>=0.0.10
langchain-chroma>=0.0.10
sentence-transformers>=2.2.2
chromadb>=0.4.6
PyYAML>=6.0
transformers>=4.30.1
torch>=2.0.0
numpy>=1.24.3
scikit-learn>=1.2.2
rouge>=1.0.1
pypdf>=3.12.1
ctransformers>=0.2.24
flask==2.3.3
requests>=2.31.0
openai>=1.3.0
azure-ai-inference>=1.0.0b9
azure-core>=1.29.4
colorama>=0.4.6
python-dotenv>=1.0.0